Setting up JWT verify as a middleware
